rename lib/ably.dart -> lib/ably_flutter_plugin.dart
