+++ lib/ably_flutter_plugin.dart
