+++ lib/src/impl/realtime/connection.dart
@@ -1,6 +1,6 @@
 import 'dart:async';
 
-import 'package:ably_flutter_plugin/ably.dart';
+import 'package:ably_flutter_plugin/ably_flutter_plugin.dart';
 import 'package:ably_flutter_plugin/src/impl/realtime/realtime.dart';
 
 import '../../spec/spec.dart' show Connection, ConnectionState, ErrorInfo;
