+++ test/ably_flutter_plugin_test.dart
@@ -1,4 +1,4 @@
-import 'package:ably_flutter_plugin/ably.dart';
+import 'package:ably_flutter_plugin/ably_flutter_plugin.dart';
 import 'package:ably_flutter_plugin/src/impl/rest/rest.dart';
 import 'package:ably_flutter_plugin/src/info.dart';
 import 'package:ably_flutter_plugin/src/platform.dart' as platform;
