+++ test_integration/lib/test/platform_and_ably_version_test.dart
@@ -1,5 +1,5 @@
 import 'package:ably_flutter_integration_test/test/test_widget_abstract.dart';
-import 'package:ably_flutter_plugin/ably.dart' as ably;
+import 'package:ably_flutter_plugin/ably_flutter_plugin.dart' as ably;
 
 import '../test_dispatcher.dart';
