+++ test_integration/lib/test/rest_publish_test.dart
@@ -1,5 +1,5 @@
 import 'package:ably_flutter_integration_test/test/test_widget_abstract.dart';
-import 'package:ably_flutter_plugin/ably.dart';
+import 'package:ably_flutter_plugin/ably_flutter_plugin.dart';
 
 import '../test_dispatcher.dart';
 import 'appkey_provision_helper.dart';
